+++ src/Convo/Core/Factory/ConvoServiceFactory.php
@@ -9,7 +9,7 @@ use Convo\Core\Intent\EntityModel;
 class ConvoServiceFactory
 {
 	const SERVICE_VERSION_ATTRIBUTE		=	'convo_service_version';
-	const SERVICE_VERSION	=	34;
+	const SERVICE_VERSION	=	35;
 
 	/**
 	 * @var \Convo\Core\Factory\PackageProviderFactory
@@ -266,6 +266,7 @@ class ConvoServiceFactory
 		$migrations[]	=	new \Convo\Core\Migrate\MigrateTo32();
 		$migrations[]	=	new \Convo\Core\Migrate\MigrateTo33();
 		$migrations[]	=	new \Convo\Core\Migrate\MigrateTo34();
+		$migrations[]	=	new \Convo\Core\Migrate\MigrateTo35();
 
 		return $migrations;
 	}